api/django_app/requirements.txt
Django==5.1.4
django-cors-headers==4.3.1
django-filter~=24.3
django-ratelimit==4.1.0
djangorestframework-simplejwt==5.3.1
djangorestframework~=3.15.1
firebase-admin~=6.6.0
google-cloud-logging==3.11.3
google_cloud_logging~=3.11.3
psutil~=5.9.8
psycopg2-binary==2.9.10 # Needed to work with Python 13.3
pydantic==2.10.5 # Needed to work with Python 13.3
pyjwt==2.8.0
python-dotenv==1.0.1
python-json-logger==2.0.7
requests==2.31.0
six==1.16.0